requirements.txt
selenium==4.17.2
selenium-wire==5.1.0
seleniumbase==4.23.2
undetected-chromedriver==3.5.5
python-dotenv>=1.0.0
lxml==4.9.3